only use Double for positions, dimensions and margins

+++ api/src/main/kotlin/io/github/mamachanko/DrawingService.kt
@@ -8,7 +8,7 @@ class DrawingService(val pageTemplates: Set<PageTemplate>, val colorPalettes: Se
     fun getDrawing(width: Double, height: Double): Drawing {
 //        TODO: randomly pick palette and strategy
         val pageTemplate = pageTemplates.first()
-        val page = Page(width = width.toInt(), height = height.toInt(), layout = pageTemplate.layout, grid = pageTemplate.grid)
+        val page = Page(width = width, height = height, layout = pageTemplate.layout, grid = pageTemplate.grid)
         return Drawing(
                 page = page,
                 palette = colorPalettes.first(),
