+++ api/src/test/kotlin/io/github/mamachanko/DrawingServiceTests.kt
@@ -15,13 +15,13 @@ class DrawingServiceTests {
         val width = 1230.45
         val height = 856.78
 
-        val pageTemplates = setOf(PageTemplate(layout = Layout(50, 50, 5), grid = Grid(2, 2)))
+        val pageTemplates = setOf(PageTemplate(layout = Layout(50.0, 50.0, 5.0), grid = Grid(2, 2)))
         val colorPalettes = setOf(RandomPalette())
         val strategies = setOf(SliceOnceStrategy())
 
         val drawing = DrawingService(pageTemplates, colorPalettes, strategies).getDrawing(width, height)
 
-        assertThat(drawing.page, `is`(equalTo(Page(width = width.toInt(), height = height.toInt(), layout = Layout(50, 50, 5), grid = Grid(2, 2)))))
+        assertThat(drawing.page, `is`(equalTo(Page(width = width, height = height, layout = Layout(50.0, 50.0, 5.0), grid = Grid(2, 2)))))
         assertThat(drawing.palette, `is`(instanceOf(RandomPalette::class.java)))
         assertThat(drawing.strategy, `is`(instanceOf(SliceOnceStrategy::class.java)))
     }
